Moved HTML css reset to begining of CSS chain
The existing css reset was at the beginning of app.css, which was after
several other CSS files were being loaded. This was overriding other
twitter boostrap styles. I added the HTML5 boilerplate reset.css file to
the beginning of the css chain, which I think should be more clean.